feat(settings): restructure into grouped sub-screens hub (#31)
