+++ src/services/i18n/locales/fr.ftl
@@ -26,6 +26,10 @@ settings-revoke-consent = Révoquer le consentement
 settings-sync-title = Synchronisation
 settings-sync-description = Appairez vos appareils sur le même réseau Wi-Fi. Les données voyagent directement entre eux, chiffrées, jamais via un serveur.
 settings-pair-device = Appairer un appareil
+settings-section-general = Général
+settings-section-intelligence = Intelligence
+settings-section-transcription = Transcription
+settings-section-privacy = Confidentialité
 
 sync-pairing-title = Appairage
 sync-show-code = Afficher un code d'appairage
