+++ src/ui/settings/intelligence.rs
@@ -1,0 +1,96 @@
+use crate::db::Database;
+use crate::services::i18n::t;
+use crate::ui::AppState;
+use dioxus::prelude::*;
+use std::sync::Arc;
+
+#[component]
+pub fn IntelligenceSettings() -> Element {
+    let db: Signal<Arc<Database>> = use_context();
+    let app: AppState = use_context();
+    let mut openai_key =
+        use_signal(|| db().get_setting("openai_api_key").unwrap_or_default());
+    let mut max_sources = use_signal(|| {
+        db().get_setting("rag_max_sources")
+            .and_then(|v| v.parse::<i64>().ok())
+            .unwrap_or(8)
+    });
+    let mut saved = use_signal(|| false);
+    let lang = (app.current_lang)();
+
+    rsx! {
+        div { class: "space-y-6 pb-20",
+            h2 { class: "text-lg font-semibold text-stone-900", {t(&lang, "settings-api-keys-title")} }
+            div {
+                label { class: "block text-sm font-medium text-stone-700 mb-1", {t(&lang, "settings-openai-label")} }
+                input {
+                    class: "w-full border border-stone-200 rounded-xl px-3 py-2.5 text-sm outline-none text-stone-900 bg-warm-white",
+                    r#type: "password",
+                    placeholder: "sk-...",
+                    value: "{openai_key}",
+                    oninput: move |evt| {
+                        openai_key.set(evt.value());
+                        saved.set(false);
+                    },
+                }
+            }
+
+            h2 { class: "text-lg font-semibold text-stone-900 pt-2", {t(&lang, "settings-search-title")} }
+            div {
+                div { class: "flex justify-between items-center mb-1",
+                    label { class: "text-sm font-medium text-stone-700",
+                        {t(&lang, "settings-max-sources")}
+                    }
+                    span { class: "text-sm font-semibold text-ios-orange",
+                        "{max_sources}"
+                    }
+                }
+                input {
+                    class: "w-full accent-ios-orange",
+                    r#type: "range",
+                    min: "3",
+                    max: "15",
+                    value: "{max_sources}",
+                    oninput: move |evt| {
+                        if let Ok(v) = evt.value().parse::<i64>() {
+                            max_sources.set(v);
+                            saved.set(false);
+                        }
+                    },
+                }
+                div { class: "flex justify-between text-xs text-stone-400",
+                    span { "3" }
+                    span { "15" }
+                }
+            }
+
+            button {
+                class: if saved() {
+                    "w-full py-2.5 rounded-xl text-sm font-medium bg-ios-green text-white"
+                } else {
+                    "w-full py-2.5 rounded-xl text-sm font-medium bg-ios-orange text-white"
+                },
+                onclick: move |_| {
+                    let ok = openai_key().trim().to_string();
+                    if !ok.is_empty() {
+                        let _ = db().set_setting("openai_api_key", &ok);
+                    }
+                    let _ = db().set_setting(
+                        "rag_max_sources",
+                        &max_sources().to_string(),
+                    );
+                    saved.set(true);
+                },
+                if saved() {
+                    {t(&lang, "settings-saved")}
+                } else {
+                    {t(&lang, "settings-save")}
+                }
+            }
+
+            p { class: "text-xs text-stone-400 text-center",
+                {t(&lang, "settings-keys-stored-locally")}
+            }
+        }
+    }
+}
