+++ src/ui/settings/transcription.rs
@@ -1,0 +1,58 @@
+use crate::db::Database;
+use crate::services::i18n::t;
+use crate::ui::AppState;
+use dioxus::prelude::*;
+use std::sync::Arc;
+
+#[component]
+pub fn TranscriptionSettings() -> Element {
+    let db: Signal<Arc<Database>> = use_context();
+    let app: AppState = use_context();
+    let mut soniox_key =
+        use_signal(|| db().get_setting("soniox_api_key").unwrap_or_default());
+    let mut saved = use_signal(|| false);
+    let lang = (app.current_lang)();
+
+    rsx! {
+        div { class: "space-y-6 pb-20",
+            h2 { class: "text-lg font-semibold text-stone-900", {t(&lang, "settings-api-keys-title")} }
+            div {
+                label { class: "block text-sm font-medium text-stone-700 mb-1", {t(&lang, "settings-soniox-label")} }
+                input {
+                    class: "w-full border border-stone-200 rounded-xl px-3 py-2.5 text-sm outline-none text-stone-900 bg-warm-white",
+                    r#type: "password",
+                    placeholder: t(&lang, "settings-soniox-placeholder"),
+                    value: "{soniox_key}",
+                    oninput: move |evt| {
+                        soniox_key.set(evt.value());
+                        saved.set(false);
+                    },
+                }
+            }
+
+            button {
+                class: if saved() {
+                    "w-full py-2.5 rounded-xl text-sm font-medium bg-ios-green text-white"
+                } else {
+                    "w-full py-2.5 rounded-xl text-sm font-medium bg-ios-orange text-white"
+                },
+                onclick: move |_| {
+                    let sk = soniox_key().trim().to_string();
+                    if !sk.is_empty() {
+                        let _ = db().set_setting("soniox_api_key", &sk);
+                    }
+                    saved.set(true);
+                },
+                if saved() {
+                    {t(&lang, "settings-saved")}
+                } else {
+                    {t(&lang, "settings-save")}
+                }
+            }
+
+            p { class: "text-xs text-stone-400 text-center",
+                {t(&lang, "settings-keys-stored-locally")}
+            }
+        }
+    }
+}
